+++ packages/dartness_server/lib/exception.dart
@@ -6,4 +6,4 @@ export 'src/exception/annotation/error_handler.dart';
 export 'src/exception/dartness_catch_handler.dart';
 export 'src/exception/dartness_error_handler.dart';
 export 'src/exception/dartness_error_handler_register.dart';
-export 'src/exception/http_status_exception.dart';
+export 'src/exception/http_status_code_exception.dart';