+++ packages/dartness_server/test/lib/src/exception/example_custom_http_status_exception.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:dartness_server/exception.dart';
 
-class ExampleCustomHttpStatusException extends HttpStatusException {
+class ExampleCustomHttpStatusException extends HttpStatusCodeException {
   const ExampleCustomHttpStatusException(String message)
       : super(message, HttpStatus.notFound);
 }